created components for logout and settings button
